Some more docs for the LNT json file format.
A few people recently have asked me how to be able to quickly test out LNT on
their own data.  For most of them, the quickest way to do so is to create a
script to translate their data format to the JSON file to be submitted to the
LNT server. The quickest way to do so is probably just seeing an example and a
simple description of what the json file structure is, rather than use the APIs
in lnt.testing.

After the evaluation of the LNT server and web-ui shows it's a useful tool, the
recommended way to produce lnt json files remains using the lnt.testing api.

Differential Revision: http://reviews.llvm.org/D19053